Fix reading distribution arrays from JSON files
The various distribution arrays were not being read/set correctly into the
corresponding table model because the distribution type was never decoded, resulting
in the type being set as junk. This cascaded all the way through the stats
generation process. Decoding the types correctly fixes the issue and allows for
a "round-trip" process.