Fixing several issues in ImportHDF5Dataset filter, widget and the filerparameter.
The Parameter took a point to a specific filter but any filter can use this type
of input
The FilterParameter class was rewritten to use the callback mechanism for getting
and setting variables in and out of the filter class
Callbacks were added to the filter parameter class
The Widget was updated to use the callback mechanism
The filter itself had its newFilterInstance() updated to a custom implementation
since the filter parameter does not have the required properties set. The PreFlightTest
will segfault when it runs this filter or if other codes would have called the
newFilterInstance() instance method.