DataArrayPath changes are detected during FilterPipeline::preflightPipeline() and each filter is updated with renamed paths before its preflight is run.  Changes are performed through each AbstractFilter's FilterParameters because not all data paths are stored as DataArrayPath or as a proxy.  The base FilterParameter class handles the generic cases for DataArrayPath and data proxies.  Subclasses have to handle specific cases where a path is stored as a QString, list of paths, etc.
This commit handles the generic cases for DataArrayPath, DataContainerArrayProxy, DataContainerProxy, AttributeMatrixProxy, and DataArrayProxy, as well as DataContainerSelectionFilterParameter, ComparisonSelectionFilterParameter, and the widgets for DataContainer/AttributeMatrix/DataArray Creation/Selection, and ComparisonSelectionWidget. NeighborLists are not accounted for because they are a special case, and a number of other FilterParameters and widgets need to be updated for it to work on every filter.