The DataStructureWidget filters out the DataArraySelectionWidgets based on if the given path meets the requirements.  Dragging a path keeps the current filtering in place until the drag is completed.  Otherwise, dragging the mouse off of any tree items would clear the filtering coming from the DataStructureWidget.
Ending a drag from the DataArrayPathSelectionWidget also clears the filtering.

DataArrayPathSelectionWidget now has custom painting to mark what type of path (DataContainer, AttributeMatrix, or DataArray) is required.  Custom icons are created for dragging to or from the DataArrayPathSelectionWidget to indicate what is being dragged.  Each DataType has its own color that is used both for the icon and for filtering the DataStructure.