Introduce new widgets. Update Style of new widgets
Created the SVSmallLabel widget that can be subclassed in QtDesigner to allow
more control over the presentation of the QLabel based class.
Update all the I/O widget codes that need to detect relative/absolute file paths
to hide/show as necessary and real-time update the path as the user types in a
value.
Implement default styles for the SVSmallLabel classes.