Python wrapping update for autogenerated files
During compile time additional python module files are generated which attempts
to wrap the C++ wrappers with a more pythonic style of pure python code. While
the files are generated the approach does not work due to the preflight and execute
issues with reusing a DataContainerArray object. A prototype test was started
near the top of the DataContainerArrayTest.py file to really dive deep into
understanding what the issues are. There are also now issues with name clashes in
the packaging and how to detect if the 'dream3d' should be named to simpl to conform
to the current packaging scheme.