Fix issue when reading a JSON file with an extended suffix.
Having more than a single '.' in the filename will cause an error when trying to
determine if the file extension is correct. The code now just uses the 'suffix'
instead of the 'completeSuffix()' function. The produced error message will also
now print the extension that was extracted from the filename.

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>