Cleanup the IconPushButton styles and remove API to get the PushButton CSS
Removed the API from QtSStyles and SVStyle to return the CSS for a QPushButton.
The applications should now be using pure style sheets instead of hard coding
the styles

Updated the four variations on the 'arrow' icons from narrow green to thicker
black which more matches the current style being used.

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>