Refactored the public variables in the *Proxy classes
+ All the public variables that were available publicly are now private and accessed
with setters/getters. Added API to 'insert' an object into the storage class in
each of the proxy classes.

+ Added comments to disable clang-formatting for the PYBIND11 macro areas as we depend
on specific formatting around the commas.
Adding in the clang on|off switches forces clang format to reformat certain files.

+ Most loops in the proxy classes were changed to range-based C++11 loops.

+ FilterParameters that use a struct class to hold data, the struct declaration and
definition was moved to its own header file. The python bindings code (pybind11)
was cleaned up due to this.

+ Additional cleanups in the template pybind11 codes to remove dependencies on OrientationLib

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>