Properly set the tuple dim values. Set the number of headers correctly
The headers were only allowed to grow so if there was a line in the header portion
of the file that had a lot of tokens, the number of columns of data would be set
to this value but never allowed to contract. The fix was to ensure an update to
the headers when the import line number was set.

The tuple dimensions were reset when reading from a pipeline file. The initial read
of the json values was correct but once placed in the table they were replaced/reset
with default number of rows further down in the initialization.

Updated some of the API to use a const QString& reference instead of making copies.

Implemented a caching algorithm for the first 50 lines of the text file and put a
QFileSystemWatcher on the file. If the file changes the cached lines will also be
updated.

updates #302. Fixes #302.

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>