Improvements to the Write ASCII Data filter. Single file mode
Added the ability for the filter to write a single file with all the data arranged
in a column format. The original mode is called 'Multi File Export' and is the
default.
Documentation is updated.
Unit Test Updated.
DREAM.3D updated to use this filter in an Example Pipeline (INL Export).

Updates #309. Fixes #309

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>