Implement proper DataArrayPath checking during preflight.
The ImportHDF5DatasetWidget was also updated to be consistent with the features
of the AbstractIOWidget where we can drag-and-drop files, show absolute paths, and
reveal the file in the operating system's file navigation(macOS Find, Windows Explorer).

AbstractIOFileWidget also received a small update to fix a logic error when verifying
that a file exists.

updates #316

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>