Fix crash in RawBinaryReader when using large SkipHeaderBytes values
The issue appeared because the assumption was that the amount of data eventually
read would be larger than the number of header bytes to skip. Implemented a UInt64
filter parameter and matching widget. This new filter parameter is used for the
RawBinaryReader so that the use can have offsets larger than 2GB (2^31). The original
value was stored in an int32_t which would eventually be inadequate.

#fixes #329. closes #329. updates #329

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>