Fix crash in RawBinaryReader when using large SkipHeaderBytes values (#330)
+ Fix crash in RawBinaryReader when using large SkipHeaderBytes values

The issue appeared because the assumption was that the amount of data eventually
read would be larger than the number of header bytes to skip. Implemented a UInt64
filter parameter and matching widget. This new filter parameter is used for the
RawBinaryReader so that the use can have offsets larger than 2GB (2^31). The original
value was stored in an int32_t which would eventually be inadequate.

+Updated API to Int64_t, Updated Copyright, Updated Unit Test, Rewrote Execute
+ Now use int64_t to pass the SkipHeaderBytes.
+ Update the copyright years in files
+ Update the Prime Contract Number
+ Rewrite the primary function that reads the data by passing the filter instance.
+ Use a switch statement to call each variation of the templated function
+ Add missing 'typename' for clang.
+ Fix GCC Compiler error  due to ambiguous cast to QVariant

#fixes #329. closes #329. updates #329

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>