Update IGeometryGrid API for set/get Dimension/Origin/Spacing
The API is changed to use a FloatVec3Type or SizeVec3Type (both alias for SIMPLArray<T, N>) so that
the API is consistent among the various classes and easy to use within SIMPL/DREAM.3D and its plugins.
The use of the std::tuple<T> based return types is deprecated and removed.

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>