Fix crash when reading prior .dream3d files with mesh based geometry
The crash is due to a bad cast from Int64 array to size_t array.
We now try to first read as the Int64 array and reinterpret cast to size_t. If that fails
then we read as a size_t (unsigned long long int) and return the result of that
operation.

fixes #363. closes #363

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>