Fix:Saving Pipelines to JSON does not expand prior pipelines.
If a "Read DREAM3D File" filter was used within a pipeline, the pipeline
is executed and then the pipeline is saved the resulting JSON file will
have the current pipeline filters in addition to any previous pipeline that
is stored in the .dream3d file. This is unexpected behavior and is fixed
with this commit.

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>