Speed up FindPointInPolyhedron by using mersenne twister for random numbers
Each loop iteration a new random generator was created that slaved out to the
random_device on the machine. On Unix and macOS this can be a slow file read
which can also become a bottle-neck during multi-threaded operations. The code
was changed to use use the mersenne-twister algorithm that is seeded through
the std::chrono clock.

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>