Check for illegal or troublesome characters when importing CSV file.
Check for &, :, / or \ as those characters can cause issues down the road when written
to disk.
Do NOT write NeighborLists to XDMF files as it doesn't make sense
Clean up FilterInputWidget in PipelineItem if it isn't nullptr.

When moving the data arrays, actually move the array.

The data arrays that were set for things such as the vertex shared list were not
being removed from the source Attribute Matrix when the user selected to have the
data MOVED instead of COPIED.

Find potential out-of-bounds access during filename parsing.

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>